FlightLogger logs PWM in 1-4.

+++ firmware/HAL_Common.X/FlightControllerCommonTypes.h
@@ -4,6 +4,7 @@
  *
  * Created on August 15, 2022, 4:14 PM
  */
+#include <stdint.h>
 
 #ifndef FLIGHTCONTROLLERCOMMONTYPES_H
 #define	FLIGHTCONTROLLERCOMMONTYPES_H
@@ -13,15 +14,15 @@ extern "C" {
 #endif
 
     struct PwmInputCapture {
-        unsigned int ch1_pulse_width_us;
-        unsigned int ch2_pulse_width_us;
-        unsigned int ch3_pulse_width_us;
-        unsigned int ch4_pulse_width_us;
+        uint16_t ch1_pulse_width_us;
+        uint16_t ch2_pulse_width_us;
+        uint16_t ch3_pulse_width_us;
+        uint16_t ch4_pulse_width_us;
     };
     typedef struct PwmInputCapture PwmInputCapture;
 
     struct PwmOutputData {
-        unsigned int pulse_ticks[5];
+        uint16_t pulse_ticks[5];
         char current_channel;
     };
     typedef struct PwmOutputData PwmOutputData;
